docs(owner-control): document scan refresh compatibility
