+++ crates/mesh-llm-host-runtime/src/protocol/mod.rs
@@ -765,1775 +765,4 @@ pub(crate) fn decode_control_frame<T: ValidateControlFrame>(
 }
 
 #[cfg(test)]
-pub(crate) mod tests {
-    use super::*;
-    use crate::crypto::OwnershipSummary;
-    use crate::mesh::{PeerInfo, resolve_peer_down, resolve_peer_leaving};
-    use crate::proto::node::{
-        ConfiguredModelRef, GossipFrame, MeshSubprotocolOpen, NodeConfigSnapshot, NodeGpuConfig,
-        NodeModelEntry, NodePluginEntry, NodeRole, OwnerControlError, OwnerControlErrorCode,
-        OwnerControlHandshake, PeerAnnouncement, RouteTableRequest, SignedNodeOwnership,
-    };
-    use iroh::{EndpointAddr, EndpointId, SecretKey};
-    use std::collections::{HashMap, HashSet};
-
-    const FULL_SURFACE_VALID_FIXTURE: &str =
-        include_str!("../../tests/fixtures/skippy_full_surface_valid.toml");
-
-    fn make_valid_gossip_frame() -> GossipFrame {
-        GossipFrame {
-            r#gen: NODE_PROTOCOL_GENERATION,
-            sender_id: vec![0u8; 32],
-            peers: vec![PeerAnnouncement {
-                endpoint_id: vec![0u8; 32],
-                role: NodeRole::Worker as i32,
-                ..Default::default()
-            }],
-        }
-    }
-
-    fn make_config_snapshot() -> NodeConfigSnapshot {
-        NodeConfigSnapshot {
-            version: 1,
-            gpu: Some(NodeGpuConfig {
-                assignment: crate::proto::node::GpuAssignment::Pinned as i32,
-            }),
-            models: vec![NodeModelEntry {
-                model: "Qwen3-8B".to_string(),
-                mmproj: Some("mmproj-cut".to_string()),
-                ctx_size: Some(8192),
-                gpu_id: Some("pci:0000:65:00.0".to_string()),
-                model_ref: Some(ConfiguredModelRef {
-                    declared_ref: "Qwen3-8B".to_string(),
-                    source_kind: None,
-                    revision: None,
-                }),
-                mmproj_ref: Some(ConfiguredModelRef {
-                    declared_ref: "mmproj-cut".to_string(),
-                    source_kind: None,
-                    revision: None,
-                }),
-            }],
-            plugins: vec![NodePluginEntry {
-                name: "demo".to_string(),
-                enabled: Some(true),
-                command: Some("mesh-llm".to_string()),
-                args: vec!["--plugin".to_string(), "demo".to_string()],
-            }],
-            config_toml: None,
-            mesh_requirements: None,
-        }
-    }
-
-    fn make_nested_mesh_config() -> crate::plugin::MeshConfig {
-        toml::from_str(
-            r#"version = 1
-
-[gpu]
-assignment = "auto"
-parallel = 2
-
-[defaults.model_fit]
-kv_unified = "auto"
-
-[defaults.hardware]
-gpu_layers = "auto"
-tensor_split = []
-
-[defaults.throughput]
-parallel = 3
-
-[defaults.skippy]
-activation_wire_dtype = "auto"
-
-[defaults.speculative]
-mode = "auto"
-
-[defaults.request_defaults]
-reasoning_budget = "auto"
-
-[defaults.multimodal]
-mmproj = "defaults-projector.gguf"
-
-[defaults.advanced.server]
-alias = "defaults-alias"
-
-[[models]]
-model = "Qwen3-8B.gguf"
-
-[models.model_fit]
-ctx_size = 16384
-
-[models.hardware]
-gpu_layers = 99
-
-[models.throughput]
-parallel = 4
-
-[models.skippy]
-binary_stage_transport = "auto"
-
-[models.speculative]
-draft_selection_policy = "auto"
-
-[models.request_defaults]
-top_p = 0.95
-
-[models.multimodal]
-mmproj = "model-projector.gguf"
-
-[models.advanced.server]
-alias = "model-alias"
-"#,
-        )
-        .expect("nested mesh config should parse")
-    }
-
-    fn make_valid_owner_control_handshake() -> OwnerControlHandshake {
-        OwnerControlHandshake {
-            ownership: Some(SignedNodeOwnership {
-                version: 1,
-                cert_id: "cert-1".to_string(),
-                owner_id: "owner-1".to_string(),
-                owner_sign_public_key: vec![0x11; 32],
-                node_endpoint_id: vec![0x22; 32],
-                issued_at_unix_ms: 1,
-                expires_at_unix_ms: 2,
-                node_label: Some("node-01".to_string()),
-                hostname_hint: Some("node-01".to_string()),
-                signature: vec![0x33; 64],
-            }),
-        }
-    }
-
-    #[test]
-    fn owner_control_handshake_empty_owner_id_uses_handshake_error() {
-        let mut handshake = make_valid_owner_control_handshake();
-        handshake
-            .ownership
-            .as_mut()
-            .expect("test handshake must include ownership")
-            .owner_id = "   ".to_string();
-
-        let err = handshake
-            .validate_frame()
-            .expect_err("handshake with blank owner_id must be rejected");
-        assert!(matches!(err, ControlFrameError::MissingControlOwnerId));
-        assert_eq!(err.to_string(), "owner control handshake missing owner_id");
-    }
-
-    #[test]
-    fn owner_control_error_rejects_invalid_error_code() {
-        for code in [OwnerControlErrorCode::Unspecified as i32, 9999] {
-            let err = OwnerControlError {
-                code,
-                message: "invalid".to_string(),
-                request_id: Some(1),
-                current_revision: None,
-            }
-            .validate_frame()
-            .expect_err("invalid owner-control error code must be rejected");
-            assert!(matches!(
-                err,
-                ControlFrameError::InvalidOwnerControlErrorCode { got } if got == code
-            ));
-            assert_eq!(
-                err.to_string(),
-                format!("invalid owner control error code: {code}")
-            );
-        }
-    }
-
-    fn make_test_peer_info(peer_id: EndpointId) -> PeerInfo {
-        PeerInfo {
-            id: peer_id,
-            addr: EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            mesh_id: None,
-            mesh_policy_hash: None,
-            genesis_policy: None,
-            role: crate::mesh::NodeRole::Worker,
-            first_joined_mesh_ts: None,
-            models: vec![],
-            vram_bytes: 0,
-            rtt_ms: None,
-            model_source: None,
-            admitted: true,
-            serving_models: vec![],
-            hosted_models: vec![],
-            hosted_models_known: false,
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec![],
-            last_seen: std::time::Instant::now(),
-            last_mentioned: std::time::Instant::now(),
-            version: None,
-            gpu_name: None,
-            hostname: None,
-            is_soc: None,
-            gpu_vram: None,
-            gpu_reserved_bytes: None,
-            gpu_mem_bandwidth_gbps: None,
-            gpu_compute_tflops_fp32: None,
-            gpu_compute_tflops_fp16: None,
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: None,
-            release_attestation_summary: crate::ReleaseAttestationSummary::default(),
-            artifact_transfer_supported: false,
-            stage_protocol_generation_supported: false,
-            stage_status_list_supported: false,
-            owner_summary: OwnershipSummary::default(),
-            advertised_model_throughput: vec![],
-
-            display_rtt: None,
-            selected_path: None,
-            propagated_latency: None,
-        }
-    }
-
-    #[test]
-    fn protocol_from_alpn_defaults_to_v1() {
-        assert_eq!(protocol_from_alpn(ALPN_V1), ControlProtocol::ProtoV1);
-        assert_eq!(
-            protocol_from_alpn(b"mesh-llm/999"),
-            ControlProtocol::ProtoV1
-        );
-    }
-    #[test]
-    fn control_frame_roundtrip() {
-        let frame = make_valid_gossip_frame();
-        let encoded = encode_control_frame(STREAM_GOSSIP, &frame);
-        let decoded: GossipFrame = decode_control_frame(STREAM_GOSSIP, &encoded)
-            .expect("valid gossip frame must decode successfully");
-        assert_eq!(decoded.r#gen, NODE_PROTOCOL_GENERATION);
-        assert_eq!(decoded.peers.len(), 1);
-        assert_eq!(decoded.peers[0].endpoint_id, vec![0u8; 32]);
-        assert_eq!(decoded.peers[0].role, NodeRole::Worker as i32);
-    }
-
-    #[test]
-    fn mesh_subprotocol_open_roundtrips_and_validates() {
-        let open = MeshSubprotocolOpen {
-            r#gen: NODE_PROTOCOL_GENERATION,
-            name: skippy_protocol::STAGE_SUBPROTOCOL_NAME.to_string(),
-            major: skippy_protocol::STAGE_SUBPROTOCOL_MAJOR,
-        };
-        let encoded = encode_control_frame(STREAM_SUBPROTOCOL, &open);
-        let decoded: MeshSubprotocolOpen =
-            decode_control_frame(STREAM_SUBPROTOCOL, &encoded).unwrap();
-        assert_eq!(decoded.name, skippy_protocol::STAGE_SUBPROTOCOL_NAME);
-        assert_eq!(decoded.major, skippy_protocol::STAGE_SUBPROTOCOL_MAJOR);
-
-        let bad = MeshSubprotocolOpen {
-            r#gen: NODE_PROTOCOL_GENERATION,
-            name: String::new(),
-            major: skippy_protocol::STAGE_SUBPROTOCOL_MAJOR,
-        };
-        let encoded = encode_control_frame(STREAM_SUBPROTOCOL, &bad);
-        let err = decode_control_frame::<MeshSubprotocolOpen>(STREAM_SUBPROTOCOL, &encoded)
-            .expect_err("empty subprotocol names must be rejected");
-        assert!(matches!(err, ControlFrameError::InvalidSubprotocol));
-    }
-
-    #[test]
-    fn proto_v1_route_table_rejects_bad_generation_or_legacy_payload() {
-        use crate::proto::node::RouteTable;
-
-        let zero_gen_req = RouteTableRequest {
-            requester_id: vec![0u8; 32],
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_ROUTE_REQUEST, &zero_gen_req);
-        let err = decode_control_frame::<RouteTableRequest>(STREAM_ROUTE_REQUEST, &encoded)
-            .expect_err("request gen=0 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 0 }),
-            "expected BadGeneration{{got:0}}, got {:?}",
-            err
-        );
-
-        let wrong_gen_req = RouteTableRequest {
-            requester_id: vec![0u8; 32],
-            r#gen: 99,
-        };
-        let encoded = encode_control_frame(STREAM_ROUTE_REQUEST, &wrong_gen_req);
-        let err = decode_control_frame::<RouteTableRequest>(STREAM_ROUTE_REQUEST, &encoded)
-            .expect_err("request gen=99 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 99 }),
-            "expected BadGeneration{{got:99}}, got {:?}",
-            err
-        );
-
-        let bad_gen_response = RouteTable {
-            entries: vec![],
-            mesh_id: None,
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_ROUTE_REQUEST, &bad_gen_response);
-        let err = decode_control_frame::<RouteTable>(STREAM_ROUTE_REQUEST, &encoded)
-            .expect_err("response gen=0 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 0 }),
-            "expected BadGeneration{{got:0}} for response, got {:?}",
-            err
-        );
-
-        let wrong_gen_response = RouteTable {
-            entries: vec![],
-            mesh_id: None,
-            r#gen: 42,
-        };
-        let encoded = encode_control_frame(STREAM_ROUTE_REQUEST, &wrong_gen_response);
-        let err = decode_control_frame::<RouteTable>(STREAM_ROUTE_REQUEST, &encoded)
-            .expect_err("response gen=42 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 42 }),
-            "expected BadGeneration{{got:42}} for response, got {:?}",
-            err
-        );
-
-        let legacy_json = b"{\"hosts\":[],\"mesh_id\":null}";
-        let mut fake_frame = vec![STREAM_ROUTE_REQUEST];
-        fake_frame.extend_from_slice(&(legacy_json.len() as u32).to_le_bytes());
-        fake_frame.extend_from_slice(legacy_json);
-        let err = decode_control_frame::<RouteTableRequest>(STREAM_ROUTE_REQUEST, &fake_frame)
-            .expect_err("legacy JSON payload must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::DecodeError(_)),
-            "expected DecodeError for JSON payload, got {:?}",
-            err
-        );
-    }
-
-    #[test]
-    fn peer_lifecycle_messages_roundtrip() {
-        use crate::proto::node::{PeerDown, PeerLeaving};
-
-        let leaving_id = EndpointId::from(SecretKey::from_bytes(&[0x55; 32]).public());
-
-        let mut peers: HashMap<EndpointId, PeerInfo> = HashMap::new();
-        peers.insert(leaving_id, make_test_peer_info(leaving_id));
-        let mut connection_ids: HashSet<EndpointId> = HashSet::new();
-        connection_ids.insert(leaving_id);
-
-        let leaving_msg = PeerLeaving {
-            peer_id: leaving_id.as_bytes().to_vec(),
-            r#gen: NODE_PROTOCOL_GENERATION,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_LEAVING, &leaving_msg);
-        let decoded_leaving: PeerLeaving = decode_control_frame(STREAM_PEER_LEAVING, &encoded)
-            .expect("valid PeerLeaving must decode");
-
-        let accepted_id = resolve_peer_leaving(leaving_id, &decoded_leaving)
-            .expect("PeerLeaving from sender itself must be accepted");
-
-        peers.remove(&accepted_id);
-        connection_ids.remove(&accepted_id);
-
-        assert!(
-            !peers.contains_key(&leaving_id),
-            "leaving peer must be removed from peers after accepted PeerLeaving"
-        );
-        assert!(
-            !connection_ids.contains(&leaving_id),
-            "leaving peer must be removed from connections after accepted PeerLeaving"
-        );
-
-        let self_id = EndpointId::from(SecretKey::from_bytes(&[0xAA; 32]).public());
-        let dead_id = EndpointId::from(SecretKey::from_bytes(&[0xBB; 32]).public());
-
-        let mut peers: HashMap<EndpointId, PeerInfo> = HashMap::new();
-        peers.insert(dead_id, make_test_peer_info(dead_id));
-        let mut connection_ids: HashSet<EndpointId> = HashSet::new();
-        connection_ids.insert(dead_id);
-
-        let down_msg = PeerDown {
-            peer_id: dead_id.as_bytes().to_vec(),
-            r#gen: NODE_PROTOCOL_GENERATION,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_DOWN, &down_msg);
-        let decoded_down: PeerDown =
-            decode_control_frame(STREAM_PEER_DOWN, &encoded).expect("valid PeerDown must decode");
-
-        let result = resolve_peer_down(self_id, dead_id, true);
-        assert_eq!(
-            result,
-            Some(dead_id),
-            "confirmed-unreachable peer must be returned for removal"
-        );
-
-        if let Some(id) = result {
-            peers.remove(&id);
-            connection_ids.remove(&id);
-        }
-
-        assert!(
-            !peers.contains_key(&dead_id),
-            "dead peer must be removed from peers when confirmed unreachable"
-        );
-        assert!(
-            !connection_ids.contains(&dead_id),
-            "dead peer must be removed from connections when confirmed unreachable"
-        );
-
-        assert_eq!(decoded_down.r#gen, NODE_PROTOCOL_GENERATION);
-    }
-
-    #[test]
-    fn peer_lifecycle_rejects_forged_sender_or_unverified_down() {
-        use crate::proto::node::{PeerDown, PeerLeaving};
-
-        let valid_peer_bytes = EndpointId::from(SecretKey::from_bytes(&[0x77; 32]).public())
-            .as_bytes()
-            .to_vec();
-
-        let bad_gen_down = PeerDown {
-            peer_id: valid_peer_bytes.clone(),
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_DOWN, &bad_gen_down);
-        let err = decode_control_frame::<PeerDown>(STREAM_PEER_DOWN, &encoded)
-            .expect_err("PeerDown gen=0 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 0 }),
-            "expected BadGeneration{{got:0}} for PeerDown, got {:?}",
-            err
-        );
-
-        let bad_gen_leaving = PeerLeaving {
-            peer_id: valid_peer_bytes.clone(),
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_LEAVING, &bad_gen_leaving);
-        let err = decode_control_frame::<PeerLeaving>(STREAM_PEER_LEAVING, &encoded)
-            .expect_err("PeerLeaving gen=0 must be rejected");
-        assert!(
-            matches!(err, ControlFrameError::BadGeneration { got: 0 }),
-            "expected BadGeneration{{got:0}} for PeerLeaving, got {:?}",
-            err
-        );
-
-        let remote_id = EndpointId::from(SecretKey::from_bytes(&[0x11; 32]).public());
-        let victim_id = EndpointId::from(SecretKey::from_bytes(&[0x22; 32]).public());
-
-        let mut peers: HashMap<EndpointId, PeerInfo> = HashMap::new();
-        peers.insert(victim_id, make_test_peer_info(victim_id));
-
-        let forged = PeerLeaving {
-            peer_id: victim_id.as_bytes().to_vec(),
-            r#gen: NODE_PROTOCOL_GENERATION,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_LEAVING, &forged);
-        let decoded: PeerLeaving = decode_control_frame(STREAM_PEER_LEAVING, &encoded)
-            .expect("structurally valid PeerLeaving must decode");
-
-        let err = resolve_peer_leaving(remote_id, &decoded)
-            .expect_err("forged PeerLeaving (peer_id != remote) must be rejected");
-        assert!(
-            matches!(err, crate::protocol::ControlFrameError::ForgedSender),
-            "expected ForgedSender, got {:?}",
-            err
-        );
-
-        assert!(
-            peers.contains_key(&victim_id),
-            "victim peer must NOT be removed when PeerLeaving is forged"
-        );
-
-        let self_id = EndpointId::from(SecretKey::from_bytes(&[0x33; 32]).public());
-        let still_alive_id = EndpointId::from(SecretKey::from_bytes(&[0x44; 32]).public());
-
-        let mut peers: HashMap<EndpointId, PeerInfo> = HashMap::new();
-        peers.insert(still_alive_id, make_test_peer_info(still_alive_id));
-
-        let result = resolve_peer_down(self_id, still_alive_id, false);
-        assert!(
-            result.is_none(),
-            "PeerDown must not trigger removal when peer is still reachable"
-        );
-
-        assert!(
-            peers.contains_key(&still_alive_id),
-            "reachable peer must NOT be removed after PeerDown with should_remove=false"
-        );
-    }
-
-    #[test]
-    fn proto_v1_control_frames_reject_legacy_json_and_wrong_gen() {
-        use crate::proto::node::{PeerDown, PeerLeaving};
-
-        // JSON bytes that look plausible for the old wire format on each stream
-        let json_gossip = b"[{\"addr\":{\"id\":\"aabbcc\",\"addrs\":[]}}]";
-        let json_tunnel_map = b"{\"owner\":\"aabbcc\",\"entries\":[]}";
-        let json_route = b"{\"hosts\":[],\"mesh_id\":null}";
-        let json_peer_down = b"\"aabbccdd\"";
-        let json_peer_leaving = b"\"aabbccdd\"";
-
-        // All migrated streams must reject legacy JSON with DecodeError
-        for (stream_type, json_bytes) in [
-            (STREAM_GOSSIP, json_gossip.as_slice()),
-            (STREAM_TUNNEL_MAP, json_tunnel_map.as_slice()),
-            (STREAM_ROUTE_REQUEST, json_route.as_slice()),
-            (STREAM_PEER_DOWN, json_peer_down.as_slice()),
-            (STREAM_PEER_LEAVING, json_peer_leaving.as_slice()),
-        ] {
-            let mut frame = vec![stream_type];
-            frame.extend_from_slice(&(json_bytes.len() as u32).to_le_bytes());
-            frame.extend_from_slice(json_bytes);
-            // Each stream uses its own message type for decode; we test gossip and route
-            // request specifically since those carry gen validation too.
-            if stream_type == STREAM_GOSSIP {
-                let err = decode_control_frame::<GossipFrame>(stream_type, &frame).expect_err(
-                    &format!("JSON must be rejected on stream {:#04x}", stream_type),
-                );
-                assert!(
-                    matches!(err, ControlFrameError::DecodeError(_)),
-                    "stream {:#04x}: expected DecodeError for JSON, got {:?}",
-                    stream_type,
-                    err
-                );
-            } else if stream_type == STREAM_ROUTE_REQUEST {
-                let err =
-                    decode_control_frame::<RouteTableRequest>(stream_type, &frame).expect_err(
-                        &format!("JSON must be rejected on stream {:#04x}", stream_type),
-                    );
-                assert!(
-                    matches!(err, ControlFrameError::DecodeError(_)),
-                    "stream {:#04x}: expected DecodeError for JSON, got {:?}",
-                    stream_type,
-                    err
-                );
-            }
-            // STREAM_TUNNEL_MAP, STREAM_PEER_DOWN, STREAM_PEER_LEAVING: JSON fails prost
-            // decode which returns DecodeError — verified via the decode_control_frame
-            // path used in the existing per-stream tests.
-        }
-
-        // All migrated streams must also reject gen=0 and gen=99 where gen is checked
-        let bad_gen_gossip = GossipFrame {
-            r#gen: 0,
-            sender_id: vec![],
-            peers: vec![PeerAnnouncement {
-                endpoint_id: vec![0u8; 32],
-                role: NodeRole::Worker as i32,
-                ..Default::default()
-            }],
-        };
-        let encoded = encode_control_frame(STREAM_GOSSIP, &bad_gen_gossip);
-        let err = decode_control_frame::<GossipFrame>(STREAM_GOSSIP, &encoded)
-            .expect_err("GossipFrame gen=0 must be rejected");
-        assert!(matches!(err, ControlFrameError::BadGeneration { got: 0 }));
-
-        let bad_gen_req = RouteTableRequest {
-            requester_id: vec![0u8; 32],
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_ROUTE_REQUEST, &bad_gen_req);
-        let err = decode_control_frame::<RouteTableRequest>(STREAM_ROUTE_REQUEST, &encoded)
-            .expect_err("RouteTableRequest gen=0 must be rejected");
-        assert!(matches!(err, ControlFrameError::BadGeneration { got: 0 }));
-
-        let bad_gen_down = PeerDown {
-            peer_id: vec![0u8; 32],
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_DOWN, &bad_gen_down);
-        let err = decode_control_frame::<PeerDown>(STREAM_PEER_DOWN, &encoded)
-            .expect_err("PeerDown gen=0 must be rejected");
-        assert!(matches!(err, ControlFrameError::BadGeneration { got: 0 }));
-
-        let bad_gen_leaving = PeerLeaving {
-            peer_id: vec![0u8; 32],
-            r#gen: 0,
-        };
-        let encoded = encode_control_frame(STREAM_PEER_LEAVING, &bad_gen_leaving);
-        let err = decode_control_frame::<PeerLeaving>(STREAM_PEER_LEAVING, &encoded)
-            .expect_err("PeerLeaving gen=0 must be rejected");
-        assert!(matches!(err, ControlFrameError::BadGeneration { got: 0 }));
-
-        // Wrong gen (e.g. 2) also rejected
-        let wrong_gen_gossip = GossipFrame {
-            r#gen: 2,
-            sender_id: vec![0u8; 32],
-            peers: vec![PeerAnnouncement {
-                endpoint_id: vec![0u8; 32],
-                role: NodeRole::Worker as i32,
-                ..Default::default()
-            }],
-        };
-        let encoded = encode_control_frame(STREAM_GOSSIP, &wrong_gen_gossip);
-        let err = decode_control_frame::<GossipFrame>(STREAM_GOSSIP, &encoded)
-            .expect_err("GossipFrame gen=2 (future version) must be rejected");
-        assert!(matches!(err, ControlFrameError::BadGeneration { got: 2 }));
-    }
-
-    #[test]
-    fn owner_fields_roundtrip_through_proto_announcement() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xAB; 32]).public());
-        let ann = super::PeerAnnouncement {
-            addr: iroh::EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            role: super::NodeRole::Worker,
-            first_joined_mesh_ts: None,
-            models: vec![],
-            vram_bytes: 0,
-            model_source: None,
-            serving_models: vec![],
-            hosted_models: None,
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec![],
-            version: None,
-            model_demand: HashMap::new(),
-            mesh_id: None,
-            mesh_policy_hash: None,
-            gpu_name: None,
-            hostname: None,
-            is_soc: None,
-            gpu_vram: None,
-            gpu_reserved_bytes: None,
-            gpu_mem_bandwidth_gbps: None,
-            gpu_compute_tflops_fp32: None,
-            gpu_compute_tflops_fp16: None,
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: Some(crate::crypto::SignedNodeOwnership {
-                claim: crate::crypto::NodeOwnershipClaim {
-                    version: 1,
-                    cert_id: "cert-123".to_string(),
-                    owner_id: "owner-abc".to_string(),
-                    owner_sign_public_key: "11".repeat(32),
-                    node_endpoint_id: "22".repeat(32),
-                    issued_at_unix_ms: 10,
-                    expires_at_unix_ms: 20,
-                    node_label: Some("studio".to_string()),
-                    hostname_hint: Some("worker-01".to_string()),
-                },
-                signature: "33".repeat(64),
-            }),
-            genesis_policy: None,
-            release_attestation: None,
-            direct_admission_proof: None,
-            artifact_transfer_supported: true,
-            stage_protocol_generation_supported: true,
-            stage_status_list_supported: true,
-            advertised_model_throughput: vec![],
-            latency_ms: None,
-            latency_source: None,
-            latency_age_ms: None,
-            latency_observer_id: None,
-        };
-        let proto_pa = local_ann_to_proto_ann(&ann);
-        let skippy = proto_pa
-            .subprotocols
-            .iter()
-            .find(|subprotocol| subprotocol.name == skippy_protocol::STAGE_SUBPROTOCOL_NAME)
-            .expect("skippy-stage subprotocol should be advertised");
-        assert_eq!(skippy.major, skippy_protocol::STAGE_SUBPROTOCOL_MAJOR);
-        assert!(
-            skippy
-                .features
-                .iter()
-                .any(|feature| feature
-                    == skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_ARTIFACT_TRANSFER)
-        );
-        assert!(
-            skippy
-                .features
-                .iter()
-                .any(|feature| feature == skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_STATUS_LIST)
-        );
-        assert!(skippy.features.iter().any(|feature| feature
-            == skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_STAGE_PROTOCOL_GENERATION_V3));
-        assert_eq!(
-            proto_pa
-                .owner_attestation
-                .as_ref()
-                .map(|att| att.owner_id.as_str()),
-            Some("owner-abc")
-        );
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert!(roundtripped.artifact_transfer_supported);
-        assert!(roundtripped.stage_status_list_supported);
-        assert!(roundtripped.stage_protocol_generation_supported);
-        let roundtripped = roundtripped
-            .owner_attestation
-            .expect("owner attestation must round-trip");
-        assert_eq!(roundtripped.claim.owner_id, "owner-abc");
-        assert_eq!(roundtripped.claim.cert_id, "cert-123");
-        assert_eq!(roundtripped.claim.node_label.as_deref(), Some("studio"));
-    }
-
-    pub(crate) fn assert_mixed_version_peer_ignores_missing_release_attestation() {
-        let proto = crate::proto::node::PeerAnnouncement {
-            endpoint_id: vec![1; 32],
-            role: crate::proto::node::NodeRole::Worker as i32,
-            version: Some("0.66.0".into()),
-            ..Default::default()
-        };
-
-        let (_addr, ann) = proto_ann_to_local(&proto).expect("announcement should decode");
-        assert!(ann.release_attestation.is_none());
-
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xBC; 32]).public());
-        let peer = crate::mesh::PeerInfo::from_announcement(
-            peer_id,
-            iroh::EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            &ann,
-            crate::crypto::OwnershipSummary::default(),
-        );
-        assert_eq!(
-            peer.release_attestation_summary.status,
-            crate::ReleaseAttestationStatus::Missing
-        );
-        assert!(!peer.release_attestation_summary.verified);
-    }
-
-    #[test]
-    fn advertised_model_throughput_roundtrips_through_proto_announcement() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xAC; 32]).public());
-        let expected_hints = vec![crate::network::metrics::ModelThroughputHint {
-            model_name: "qwen".to_string(),
-            avg_tokens_per_second_milli: 42_000,
-            throughput_samples: 7,
-        }];
-        let ann = super::PeerAnnouncement {
-            addr: iroh::EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            role: super::NodeRole::Host { http_port: 9337 },
-            first_joined_mesh_ts: None,
-            models: vec![],
-            vram_bytes: 0,
-            model_source: None,
-            serving_models: vec!["qwen".to_string()],
-            hosted_models: Some(vec!["qwen".to_string()]),
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec![],
-            version: None,
-            model_demand: HashMap::new(),
-            mesh_id: None,
-            mesh_policy_hash: None,
-            gpu_name: None,
-            hostname: None,
-            is_soc: None,
-            gpu_vram: None,
-            gpu_reserved_bytes: None,
-            gpu_mem_bandwidth_gbps: None,
-            gpu_compute_tflops_fp32: None,
-            gpu_compute_tflops_fp16: None,
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: None,
-            genesis_policy: None,
-            release_attestation: None,
-            direct_admission_proof: None,
-            artifact_transfer_supported: false,
-            stage_protocol_generation_supported: false,
-            stage_status_list_supported: false,
-            advertised_model_throughput: vec![
-                expected_hints[0].clone(),
-                crate::network::metrics::ModelThroughputHint {
-                    model_name: "ghost".to_string(),
-                    avg_tokens_per_second_milli: 250_000,
-                    throughput_samples: 99,
-                },
-            ],
-            latency_ms: None,
-            latency_source: None,
-            latency_age_ms: None,
-            latency_observer_id: None,
-        };
-
-        let mut proto_pa = local_ann_to_proto_ann(&ann);
-        assert_eq!(proto_pa.advertised_model_throughput.len(), 1);
-        assert_eq!(proto_pa.advertised_model_throughput[0].model_name, "qwen");
-        assert_eq!(
-            proto_pa.advertised_model_throughput[0].avg_tokens_per_second_milli,
-            42_000
-        );
-        assert_eq!(
-            proto_pa.advertised_model_throughput[0].throughput_samples,
-            7
-        );
-        proto_pa
-            .advertised_model_throughput
-            .push(crate::proto::node::AdvertisedModelThroughput {
-                model_name: "ghost".to_string(),
-                avg_tokens_per_second_milli: 250_000,
-                throughput_samples: 99,
-            });
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(roundtripped.advertised_model_throughput, expected_hints);
-    }
-
-    #[test]
-    fn proto_announcement_without_current_stage_generation_is_not_stage_compatible() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xCD; 32]).public());
-        let proto_pa = crate::proto::node::PeerAnnouncement {
-            endpoint_id: peer_id.as_bytes().to_vec(),
-            role: crate::proto::node::NodeRole::Worker as i32,
-            subprotocols: vec![crate::proto::node::MeshSubprotocol {
-                name: skippy_protocol::STAGE_SUBPROTOCOL_NAME.to_string(),
-                major: skippy_protocol::STAGE_SUBPROTOCOL_MAJOR,
-                features: vec![
-                    skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_STAGE_CONTROL.to_string(),
-                    skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_STATUS_LIST.to_string(),
-                ],
-            }],
-            ..Default::default()
-        };
-
-        let (_, ann) = proto_ann_to_local(&proto_pa).expect("proto announcement should decode");
-
-        assert!(!ann.stage_protocol_generation_supported);
-        assert!(ann.stage_status_list_supported);
-    }
-
-    #[test]
-    fn proto_announcement_without_stage_control_is_not_stage_compatible() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xCE; 32]).public());
-        let proto_pa = crate::proto::node::PeerAnnouncement {
-            endpoint_id: peer_id.as_bytes().to_vec(),
-            role: crate::proto::node::NodeRole::Worker as i32,
-            subprotocols: vec![crate::proto::node::MeshSubprotocol {
-                name: skippy_protocol::STAGE_SUBPROTOCOL_NAME.to_string(),
-                major: skippy_protocol::STAGE_SUBPROTOCOL_MAJOR,
-                features: vec![
-                    skippy_protocol::STAGE_SUBPROTOCOL_FEATURE_STAGE_PROTOCOL_GENERATION_V3
-                        .to_string(),
-                ],
-            }],
-            ..Default::default()
-        };
-
-        let (_, ann) = proto_ann_to_local(&proto_pa).expect("proto announcement should decode");
-
-        assert!(!ann.stage_protocol_generation_supported);
-    }
-
-    #[test]
-    fn test_proto_round_trip_with_bandwidth_and_tflops() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xBC; 32]).public());
-        let ann = super::PeerAnnouncement {
-            addr: EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            role: super::NodeRole::Host { http_port: 3131 },
-            first_joined_mesh_ts: None,
-            models: vec!["Qwen".to_string()],
-            vram_bytes: 48_000_000_000,
-            model_source: Some("Qwen.gguf".to_string()),
-            serving_models: vec!["Qwen".to_string()],
-            hosted_models: Some(vec!["Qwen".to_string()]),
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec!["Qwen/Qwen3-Coder-Next-GGUF@main:Q4_K_M".to_string()],
-            version: Some("0.52.0".to_string()),
-            model_demand: HashMap::new(),
-            mesh_id: Some("mesh-proto-roundtrip".to_string()),
-            mesh_policy_hash: None,
-            gpu_name: Some("NVIDIA A100".to_string()),
-            hostname: Some("worker-01".to_string()),
-            is_soc: Some(false),
-            gpu_vram: Some("51539607552".to_string()),
-            gpu_reserved_bytes: Some("1073741824".to_string()),
-            gpu_mem_bandwidth_gbps: Some("1948.70".to_string()),
-            gpu_compute_tflops_fp32: Some("19.50".to_string()),
-            gpu_compute_tflops_fp16: Some("312.00".to_string()),
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: None,
-            genesis_policy: None,
-            release_attestation: None,
-            direct_admission_proof: None,
-            artifact_transfer_supported: true,
-            stage_protocol_generation_supported: true,
-            stage_status_list_supported: true,
-            advertised_model_throughput: vec![],
-            latency_ms: None,
-            latency_source: None,
-            latency_age_ms: None,
-            latency_observer_id: None,
-        };
-
-        let proto_pa = local_ann_to_proto_ann(&ann);
-        let hardware = proto_pa
-            .hardware
-            .as_ref()
-            .expect("hardware info must be present");
-        assert_eq!(hardware.hostname.as_deref(), Some("worker-01"));
-        assert_eq!(hardware.is_soc, Some(false));
-        assert_eq!(hardware.gpus.len(), 1);
-        assert_eq!(hardware.gpus[0].name.as_deref(), Some("NVIDIA A100"));
-        assert_eq!(hardware.gpus[0].vram_bytes.as_deref(), Some("51539607552"));
-        assert_eq!(
-            hardware.gpus[0].reserved_bytes.as_deref(),
-            Some("1073741824")
-        );
-        assert_eq!(
-            hardware.gpus[0].mem_bandwidth_gbps.as_deref(),
-            Some("1948.70")
-        );
-        assert_eq!(
-            hardware.gpus[0].compute_tflops_fp32.as_deref(),
-            Some("19.50")
-        );
-        assert_eq!(
-            hardware.gpus[0].compute_tflops_fp16.as_deref(),
-            Some("312.00")
-        );
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(
-            roundtripped.gpu_reserved_bytes.as_deref(),
-            Some("1073741824")
-        );
-        assert_eq!(
-            roundtripped.gpu_mem_bandwidth_gbps.as_deref(),
-            Some("1948.70")
-        );
-        assert_eq!(
-            roundtripped.gpu_compute_tflops_fp32.as_deref(),
-            Some("19.50")
-        );
-        assert_eq!(
-            roundtripped.gpu_compute_tflops_fp16.as_deref(),
-            Some("312.00")
-        );
-        assert_eq!(
-            roundtripped.explicit_model_interests,
-            vec!["Qwen/Qwen3-Coder-Next-GGUF@main:Q4_K_M".to_string()]
-        );
-    }
-
-    #[test]
-    fn test_proto_backward_compat_missing_tflops() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xCD; 32]).public());
-        let proto_pa = crate::proto::node::PeerAnnouncement {
-            endpoint_id: peer_id.as_bytes().to_vec(),
-            role: NodeRole::Worker as i32,
-            gpu_name: Some("NVIDIA A100".to_string()),
-            gpu_vram: Some("51539607552".to_string()),
-            hardware: Some(crate::proto::node::HardwareInfo {
-                is_soc: Some(false),
-                hostname: None,
-                gpus: vec![crate::proto::node::GpuInfo {
-                    name: Some("NVIDIA A100".to_string()),
-                    vram_bytes: Some("51539607552".to_string()),
-                    reserved_bytes: None,
-                    mem_bandwidth_gbps: Some("1948.70".to_string()),
-                    compute_tflops_fp32: None,
-                    compute_tflops_fp16: None,
-                }],
-            }),
-            ..Default::default()
-        };
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(roundtripped.gpu_reserved_bytes, None);
-        assert_eq!(
-            roundtripped.gpu_mem_bandwidth_gbps.as_deref(),
-            Some("1948.70")
-        );
-        assert_eq!(roundtripped.gpu_compute_tflops_fp32, None);
-        assert_eq!(roundtripped.gpu_compute_tflops_fp16, None);
-    }
-
-    #[test]
-    fn test_proto_gpu_info_preserves_legacy_fields_for_old_consumers() {
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xCE; 32]).public());
-        let proto_pa = crate::proto::node::PeerAnnouncement {
-            endpoint_id: peer_id.as_bytes().to_vec(),
-            role: NodeRole::Worker as i32,
-            hardware: Some(crate::proto::node::HardwareInfo {
-                is_soc: Some(false),
-                hostname: Some("worker-01".to_string()),
-                gpus: vec![
-                    crate::proto::node::GpuInfo {
-                        name: Some("NVIDIA A100".to_string()),
-                        vram_bytes: Some("51539607552".to_string()),
-                        reserved_bytes: Some("1073741824".to_string()),
-                        mem_bandwidth_gbps: Some("1948.70".to_string()),
-                        compute_tflops_fp32: Some("19.50".to_string()),
-                        compute_tflops_fp16: Some("312.00".to_string()),
-                    },
-                    crate::proto::node::GpuInfo {
-                        name: Some("NVIDIA A100".to_string()),
-                        vram_bytes: Some("51539607552".to_string()),
-                        reserved_bytes: None,
-                        mem_bandwidth_gbps: Some("1948.70".to_string()),
-                        compute_tflops_fp32: Some("19.50".to_string()),
-                        compute_tflops_fp16: Some("312.00".to_string()),
-                    },
-                ],
-            }),
-            ..Default::default()
-        };
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(roundtripped.hostname.as_deref(), Some("worker-01"));
-        assert_eq!(roundtripped.gpu_name.as_deref(), Some("2× NVIDIA A100"));
-        assert_eq!(
-            roundtripped.gpu_vram.as_deref(),
-            Some("51539607552,51539607552")
-        );
-        assert_eq!(
-            roundtripped.gpu_reserved_bytes.as_deref(),
-            Some("1073741824,")
-        );
-        assert_eq!(
-            roundtripped.gpu_mem_bandwidth_gbps.as_deref(),
-            Some("1948.70,1948.70")
-        );
-        assert_eq!(
-            roundtripped.gpu_compute_tflops_fp32.as_deref(),
-            Some("19.50,19.50")
-        );
-        assert_eq!(
-            roundtripped.gpu_compute_tflops_fp16.as_deref(),
-            Some("312.00,312.00")
-        );
-        assert_eq!(roundtripped.is_soc, Some(false));
-    }
-
-    #[test]
-    fn mesh_config_proto_roundtrip() {
-        let snapshot = make_config_snapshot();
-        let config = proto_config_to_mesh(&snapshot);
-        assert_mesh_config_from_proto(&config);
-
-        let roundtripped = mesh_config_to_proto(&config);
-        assert_proto_config_roundtrip_matches(&roundtripped, &snapshot);
-    }
-
-    fn assert_mesh_config_from_proto(config: &crate::plugin::MeshConfig) {
-        assert_eq!(config.version, Some(1));
-        assert_eq!(config.gpu.assignment, crate::plugin::GpuAssignment::Pinned);
-        assert_eq!(config.models.len(), 1);
-        assert_eq!(config.models[0].model, "Qwen3-8B");
-        assert_eq!(config.models[0].mmproj.as_deref(), Some("mmproj-cut"));
-        assert_eq!(config.models[0].ctx_size, Some(8192));
-        assert_eq!(config.models[0].gpu_id.as_deref(), Some("pci:0000:65:00.0"));
-        assert_eq!(config.plugins.len(), 1);
-        assert_eq!(config.plugins[0].name, "demo");
-    }
-
-    fn assert_proto_config_roundtrip_matches(
-        roundtripped: &NodeConfigSnapshot,
-        snapshot: &NodeConfigSnapshot,
-    ) {
-        assert_eq!(roundtripped.version, snapshot.version);
-        assert_eq!(
-            roundtripped.gpu.as_ref().map(|g| g.assignment),
-            Some(crate::proto::node::GpuAssignment::Pinned as i32)
-        );
-        assert_eq!(roundtripped.models.len(), snapshot.models.len());
-        assert_eq!(roundtripped.models[0].model, snapshot.models[0].model);
-        assert_eq!(roundtripped.models[0].mmproj, snapshot.models[0].mmproj);
-        assert_eq!(roundtripped.models[0].ctx_size, snapshot.models[0].ctx_size);
-        assert_eq!(roundtripped.models[0].gpu_id, snapshot.models[0].gpu_id);
-        assert_eq!(
-            roundtripped.models[0].model_ref,
-            snapshot.models[0].model_ref
-        );
-        assert_eq!(
-            roundtripped.models[0].mmproj_ref,
-            snapshot.models[0].mmproj_ref
-        );
-        assert_eq!(roundtripped.plugins.len(), snapshot.plugins.len());
-        assert_eq!(roundtripped.plugins[0].name, snapshot.plugins[0].name);
-        assert!(
-            roundtripped
-                .config_toml
-                .as_deref()
-                .is_some_and(|toml| toml.contains("model = \"Qwen3-8B\"")),
-            "re-encoded snapshots should include canonical config_toml payload"
-        );
-    }
-
-    #[test]
-    fn mesh_config_proto_roundtrip_preserves_nested_sections() {
-        let config = make_nested_mesh_config();
-
-        let snapshot = mesh_config_to_proto(&config);
-        let restored = proto_config_to_mesh(&snapshot);
-
-        let json = serde_json::to_value(&restored).expect("restored config should serialize");
-        assert_eq!(json["defaults"]["model_fit"]["kv_unified"], "auto");
-        assert_eq!(json["defaults"]["hardware"]["gpu_layers"], "auto");
-        assert_eq!(json["defaults"]["throughput"]["parallel"], 3);
-        assert_eq!(json["defaults"]["skippy"]["activation_wire_dtype"], "auto");
-        assert_eq!(json["defaults"]["speculative"]["mode"], "auto");
-        assert_eq!(
-            json["defaults"]["request_defaults"]["reasoning_budget"],
-            "auto"
-        );
-        assert_eq!(
-            json["defaults"]["multimodal"]["mmproj"],
-            "defaults-projector.gguf"
-        );
-        assert_eq!(
-            json["defaults"]["advanced"]["server"]["alias"],
-            "defaults-alias"
-        );
-
-        assert_eq!(json["models"][0]["model_fit"]["ctx_size"], 16384);
-        assert_eq!(json["models"][0]["hardware"]["gpu_layers"], 99);
-        assert_eq!(json["models"][0]["throughput"]["parallel"], 4);
-        assert_eq!(
-            json["models"][0]["skippy"]["binary_stage_transport"],
-            "auto"
-        );
-        assert_eq!(
-            json["models"][0]["speculative"]["draft_selection_policy"],
-            "auto"
-        );
-        assert_eq!(json["models"][0]["request_defaults"]["top_p"], 0.95);
-        assert_eq!(
-            json["models"][0]["multimodal"]["mmproj"],
-            "model-projector.gguf"
-        );
-        assert_eq!(
-            json["models"][0]["advanced"]["server"]["alias"],
-            "model-alias"
-        );
-    }
-
-    #[test]
-    fn mesh_config_proto_invalid_full_payload_falls_back_to_legacy_fields() {
-        let mut snapshot = make_config_snapshot();
-        snapshot.config_toml = Some("not valid toml = [".to_string());
-
-        let restored = proto_config_to_mesh(&snapshot);
-
-        assert_eq!(restored.models[0].model, "Qwen3-8B");
-        assert_eq!(restored.models[0].ctx_size, Some(8192));
-        assert!(restored.defaults.is_none());
-    }
-
-    #[test]
-    fn mesh_config_proto_strict_invalid_full_payload_is_rejected() {
-        let mut snapshot = make_config_snapshot();
-        snapshot.config_toml = Some("not valid toml = [".to_string());
-
-        let err = proto_config_to_mesh_strict(&snapshot).unwrap_err();
-
-        assert!(err.to_string().contains("invalid full config_toml payload"));
-    }
-
-    #[test]
-    fn mesh_config_proto_strict_legacy_payload_still_restores_fields() {
-        let mut snapshot = make_config_snapshot();
-        snapshot.config_toml = None;
-
-        let restored = proto_config_to_mesh_strict(&snapshot).unwrap();
-
-        assert_eq!(restored.models[0].model, "Qwen3-8B");
-        assert_eq!(restored.models[0].ctx_size, Some(8192));
-    }
-
-    #[test]
-    fn config_sync_prefers_structured_model_refs() {
-        let snapshot = NodeConfigSnapshot {
-            version: 1,
-            gpu: Some(NodeGpuConfig {
-                assignment: crate::proto::node::GpuAssignment::Auto as i32,
-            }),
-            models: vec![NodeModelEntry {
-                model: "legacy.gguf".to_string(),
-                mmproj: Some("legacy-mmproj.gguf".to_string()),
-                ctx_size: Some(4096),
-                gpu_id: None,
-                model_ref: Some(ConfiguredModelRef {
-                    declared_ref: "structured.gguf".to_string(),
-                    source_kind: Some("huggingface".to_string()),
-                    revision: Some("main".to_string()),
-                }),
-                mmproj_ref: Some(ConfiguredModelRef {
-                    declared_ref: "structured-mmproj.gguf".to_string(),
-                    source_kind: Some("huggingface".to_string()),
-                    revision: Some("main".to_string()),
-                }),
-            }],
-            plugins: vec![],
-            config_toml: None,
-            mesh_requirements: None,
-        };
-
-        let restored = proto_config_to_mesh(&snapshot);
-
-        assert_eq!(restored.models[0].model, "structured.gguf");
-        assert_eq!(
-            restored.models[0].mmproj.as_deref(),
-            Some("structured-mmproj.gguf")
-        );
-    }
-
-    #[test]
-    fn config_sync_empty_structured_refs_fall_back_to_legacy_strings() {
-        let snapshot = NodeConfigSnapshot {
-            version: 1,
-            gpu: Some(NodeGpuConfig {
-                assignment: crate::proto::node::GpuAssignment::Auto as i32,
-            }),
-            models: vec![NodeModelEntry {
-                model: "legacy.gguf".to_string(),
-                mmproj: Some("legacy-mmproj.gguf".to_string()),
-                ctx_size: None,
-                gpu_id: None,
-                model_ref: Some(ConfiguredModelRef {
-                    declared_ref: "   ".to_string(),
-                    source_kind: Some("huggingface".to_string()),
-                    revision: Some("main".to_string()),
-                }),
-                mmproj_ref: Some(ConfiguredModelRef {
-                    declared_ref: "".to_string(),
-                    source_kind: Some("huggingface".to_string()),
-                    revision: Some("main".to_string()),
-                }),
-            }],
-            plugins: vec![],
-            config_toml: None,
-            mesh_requirements: None,
-        };
-
-        let restored = proto_config_to_mesh(&snapshot);
-
-        assert_eq!(restored.models[0].model, "legacy.gguf");
-        assert_eq!(
-            restored.models[0].mmproj.as_deref(),
-            Some("legacy-mmproj.gguf")
-        );
-    }
-
-    #[test]
-    fn canonical_config_hash_is_stable() {
-        let snapshot = make_config_snapshot();
-        let hash1 = canonical_config_hash(&snapshot);
-        let hash2 = canonical_config_hash(&snapshot);
-        assert_eq!(hash1, hash2, "same config must produce the same hash");
-        assert_eq!(hash1.len(), 32);
-
-        let mut different = snapshot.clone();
-        different.version = 2;
-        let hash3 = canonical_config_hash(&different);
-        assert_ne!(hash1, hash3, "different config must produce different hash");
-    }
-
-    #[test]
-    fn canonical_config_hash_changes_when_structured_refs_change_encoding() {
-        let mut legacy_only = make_config_snapshot();
-        legacy_only.models[0].model_ref = None;
-        legacy_only.models[0].mmproj_ref = None;
-
-        let dual_encoded = make_config_snapshot();
-
-        assert_ne!(
-            canonical_config_hash(&legacy_only),
-            canonical_config_hash(&dual_encoded),
-            "legacy-only and dual-encoded snapshots currently have distinct hashes"
-        );
-    }
-    #[test]
-    pub(crate) fn mesh_requirements_survive_owner_control_config_round_trip() {
-        // Regression: NodeConfigSnapshot used to drop [mesh_requirements] on the
-        // owner-control get/apply path, silently stripping admission requirements
-        // from an immutable mesh. The proto NodeConfigSnapshot now carries an
-        // additive `mesh_requirements` field that mesh_config_to_proto and
-        // proto_config_to_mesh round-trip end-to-end.
-        use crate::plugin::{MeshRequirementsConfig, OwnerControlConfig};
-        let original = crate::plugin::MeshConfig {
-            version: Some(1),
-            gpu: Default::default(),
-            mesh_requirements: MeshRequirementsConfig {
-                min_node_version: Some("0.65.0".to_string()),
-                max_node_version: Some("0.66.0".to_string()),
-                min_protocol_version: Some(1),
-                max_protocol_version: Some(3),
-                require_release_attestation: true,
-                release_signer_keys: vec![
-                    "ed25519:d75a980182b10ab7d54bfed3c964073a0ee172f3daa62325af021a68f707511a"
-                        .to_string(),
-                    "ed25519:3d4017c3e843895a92b70aa74d1b7ebc9c982ccf2ec4968cc0cd55f12af4660c"
-                        .to_string(),
-                ],
-            },
-            owner_control: OwnerControlConfig::default(),
-            telemetry: Default::default(),
-            defaults: None,
-            runtime: Default::default(),
-            models: vec![],
-            plugins: vec![],
-            extra: Default::default(),
-        };
-        let snapshot = mesh_config_to_proto(&original);
-        assert!(
-            snapshot.mesh_requirements.is_some(),
-            "non-default mesh_requirements must serialize to the proto snapshot"
-        );
-        let restored = proto_config_to_mesh(&snapshot);
-        assert_eq!(
-            restored.mesh_requirements, original.mesh_requirements,
-            "mesh_requirements must round-trip through owner-control config get/apply"
-        );
-
-        // Default mesh_requirements should remain omitted on the wire so older
-        // peers continue to round-trip with absent field semantics.
-        let default_only = crate::plugin::MeshConfig::default();
-        let default_snapshot = mesh_config_to_proto(&default_only);
-        assert!(
-            default_snapshot.mesh_requirements.is_none(),
-            "default mesh_requirements must not be encoded on the wire"
-        );
-        let default_restored = proto_config_to_mesh(&default_snapshot);
-        assert_eq!(
-            default_restored.mesh_requirements,
-            crate::plugin::MeshRequirementsConfig::default()
-        );
-    }
-
-    #[test]
-    fn config_sync_empty_config_roundtrip() {
-        let config = crate::plugin::MeshConfig::default();
-        let snapshot = mesh_config_to_proto(&config);
-        let restored = proto_config_to_mesh(&snapshot);
-        assert!(restored.models.is_empty());
-        assert!(restored.plugins.is_empty());
-    }
-
-    #[test]
-    fn config_sync_config_toml_roundtrips_additive_defaults_sections() {
-        use crate::plugin::{
-            ModelConfigDefaults, ModelFitConfig, RequestDefaultsConfig, ThroughputConfig,
-        };
-        let config = crate::plugin::MeshConfig {
-            version: Some(1),
-            defaults: Some(ModelConfigDefaults {
-                throughput: Some(ThroughputConfig {
-                    parallel: Some(6),
-                    ..Default::default()
-                }),
-                model_fit: Some(ModelFitConfig {
-                    flash_attention: Some(skippy_protocol::FlashAttentionType::Disabled),
-                    ..Default::default()
-                }),
-                request_defaults: Some(RequestDefaultsConfig {
-                    reasoning_format: Some("deepseek".to_string()),
-                    ..Default::default()
-                }),
-                ..Default::default()
-            }),
-            ..Default::default()
-        };
-
-        let snapshot = mesh_config_to_proto(&config);
-        let config_toml = snapshot
-            .config_toml
-            .as_deref()
-            .expect("config TOML should serialize");
-        assert!(
-            config_toml.contains("parallel") && config_toml.contains("reasoning_format"),
-            "config TOML should carry additive defaults values: {config_toml}"
-        );
-
-        let restored = proto_config_to_mesh(&snapshot);
-        assert_eq!(
-            restored
-                .extra
-                .get("defaults")
-                .and_then(|defaults| defaults.get("throughput"))
-                .and_then(|throughput| throughput.get("parallel"))
-                .and_then(toml::Value::as_integer)
-                .or_else(|| {
-                    restored
-                        .defaults
-                        .as_ref()
-                        .and_then(|defaults| defaults.throughput.as_ref())
-                        .and_then(|throughput| throughput.parallel)
-                        .map(|parallel| parallel as i64)
-                }),
-            Some(6)
-        );
-        assert_eq!(
-            restored
-                .extra
-                .get("defaults")
-                .and_then(|defaults| defaults.get("request_defaults"))
-                .and_then(|request_defaults| request_defaults.get("reasoning_format"))
-                .and_then(toml::Value::as_str)
-                .or_else(|| {
-                    restored
-                        .defaults
-                        .as_ref()
-                        .and_then(|defaults| defaults.request_defaults.as_ref())
-                        .and_then(|request_defaults| request_defaults.reasoning_format.as_deref())
-                }),
-            Some("deepseek")
-        );
-    }
-
-    #[test]
-    fn config_sync_config_hash_determinism() {
-        use crate::plugin::{GpuAssignment, GpuConfig, ModelConfigEntry};
-        let config = crate::plugin::MeshConfig {
-            version: Some(1),
-            gpu: GpuConfig {
-                assignment: GpuAssignment::Auto,
-                parallel: None,
-            },
-            mesh_requirements: Default::default(),
-            owner_control: Default::default(),
-            telemetry: Default::default(),
-            defaults: None,
-            runtime: Default::default(),
-            models: vec![ModelConfigEntry {
-                model: "test.gguf".to_string(),
-                mmproj: None,
-                ctx_size: None,
-                gpu_id: None,
-                parallel: None,
-                cache_type_k: None,
-                cache_type_v: None,
-                batch: None,
-                ubatch: None,
-                flash_attention: None,
-                ..Default::default()
-            }],
-            plugins: vec![],
-            extra: Default::default(),
-        };
-        let snap1 = mesh_config_to_proto(&config);
-        let snap2 = mesh_config_to_proto(&config);
-        let h1 = canonical_config_hash(&snap1);
-        let h2 = canonical_config_hash(&snap2);
-        assert_eq!(h1, h2, "same config must produce same hash");
-
-        let config2 = crate::plugin::MeshConfig {
-            version: Some(1),
-            gpu: GpuConfig {
-                assignment: GpuAssignment::Auto,
-                parallel: None,
-            },
-            mesh_requirements: Default::default(),
-            owner_control: Default::default(),
-            telemetry: Default::default(),
-            defaults: None,
-            runtime: Default::default(),
-            models: vec![ModelConfigEntry {
-                model: "other.gguf".to_string(),
-                mmproj: None,
-                ctx_size: None,
-                gpu_id: None,
-                parallel: None,
-                cache_type_k: None,
-                cache_type_v: None,
-                batch: None,
-                ubatch: None,
-                flash_attention: None,
-                ..Default::default()
-            }],
-            plugins: vec![],
-            extra: Default::default(),
-        };
-        let snap3 = mesh_config_to_proto(&config2);
-        let h3 = canonical_config_hash(&snap3);
-        assert_ne!(h1, h3, "different config must produce different hash");
-    }
-
-    #[test]
-    fn mesh_config_proto_roundtrip_preserves_integrated_fixture_and_owner_control_toml() {
-        let config: crate::plugin::MeshConfig = toml::from_str(FULL_SURFACE_VALID_FIXTURE).unwrap();
-        let snapshot = mesh_config_to_proto(&config);
-
-        assert!(
-            snapshot
-                .config_toml
-                .as_deref()
-                .is_some_and(|toml| toml.contains("prefill_chunk_schedule = \"128,256,384\""))
-        );
-
-        let restored = proto_config_to_mesh(&snapshot);
-        let json = serde_json::to_value(&restored).expect("restored config serializes");
-        assert_eq!(json["owner_control"]["bind"], "127.0.0.1:7447");
-        assert_eq!(
-            json["defaults"]["request_defaults"]["reasoning_budget"],
-            256
-        );
-        assert_eq!(json["models"][0]["hardware"]["stage_layer_start"], 12);
-        assert_eq!(
-            json["models"][0]["skippy"]["prefill_chunk_schedule"],
-            "128,256,384"
-        );
-        assert_eq!(json["models"][0]["speculative"]["draft_gpu_layers"], 12);
-        assert_eq!(
-            json["models"][1]["hardware"]["model_path"],
-            "/models/gemma.gguf"
-        );
-    }
-
-    #[test]
-    fn pinned_gpu_proto_roundtrip() {
-        use crate::plugin::{GpuAssignment, GpuConfig, ModelConfigEntry};
-
-        let config = crate::plugin::MeshConfig {
-            version: Some(1),
-            gpu: GpuConfig {
-                assignment: GpuAssignment::Pinned,
-                parallel: None,
-            },
-            mesh_requirements: Default::default(),
-            owner_control: Default::default(),
-            telemetry: Default::default(),
-            defaults: None,
-            runtime: Default::default(),
-            models: vec![ModelConfigEntry {
-                model: "Qwen3-8B-Q4_K_M".to_string(),
-                mmproj: Some("mmproj-f16.gguf".to_string()),
-                ctx_size: Some(8192),
-                gpu_id: Some("pci:0000:65:00.0".to_string()),
-                parallel: None,
-                cache_type_k: None,
-                cache_type_v: None,
-                batch: None,
-                ubatch: None,
-                flash_attention: None,
-                ..Default::default()
-            }],
-            plugins: vec![],
-            extra: Default::default(),
-        };
-
-        let snapshot = mesh_config_to_proto(&config);
-        assert_eq!(
-            snapshot.gpu.as_ref().map(|gpu| gpu.assignment),
-            Some(crate::proto::node::GpuAssignment::Pinned as i32),
-            "pinned snapshots must not be downgraded to auto"
-        );
-        assert_eq!(
-            snapshot.models[0].gpu_id.as_deref(),
-            Some("pci:0000:65:00.0"),
-            "proto snapshot must carry per-model gpu_id"
-        );
-
-        let restored = proto_config_to_mesh(&snapshot);
-        assert_eq!(restored.gpu.assignment, GpuAssignment::Pinned);
-        assert_eq!(
-            restored.models[0].gpu_id.as_deref(),
-            Some("pci:0000:65:00.0")
-        );
-
-        let roundtripped = mesh_config_to_proto(&restored);
-        assert_eq!(
-            roundtripped.gpu.as_ref().map(|gpu| gpu.assignment),
-            Some(crate::proto::node::GpuAssignment::Pinned as i32),
-            "re-encoded snapshot must keep pinned assignment"
-        );
-        assert_eq!(
-            roundtripped.models[0].gpu_id.as_deref(),
-            Some("pci:0000:65:00.0"),
-            "re-encoded snapshot must keep gpu_id presence and value"
-        );
-    }
-
-    #[test]
-    fn pinned_gpu_proto_hash_changes_when_gpu_id_changes() {
-        let mut snapshot_a = make_config_snapshot();
-        snapshot_a.models[0].gpu_id = Some("pci:0000:65:00.0".to_string());
-
-        let mut snapshot_b = snapshot_a.clone();
-        snapshot_b.models[0].gpu_id = Some("pci:0000:66:00.0".to_string());
-
-        assert_ne!(
-            canonical_config_hash(&snapshot_a),
-            canonical_config_hash(&snapshot_b),
-            "changing only gpu_id must change the canonical config hash"
-        );
-    }
-
-    #[test]
-    fn pinned_gpu_proto_missing_gpu_id_decodes_as_none() {
-        let snapshot = NodeConfigSnapshot {
-            version: 1,
-            gpu: Some(NodeGpuConfig {
-                assignment: crate::proto::node::GpuAssignment::Pinned as i32,
-            }),
-            models: vec![NodeModelEntry {
-                model: "Qwen3-8B-Q4_K_M".to_string(),
-                mmproj: None,
-                ctx_size: Some(4096),
-                gpu_id: None,
-                model_ref: Some(ConfiguredModelRef {
-                    declared_ref: "Qwen3-8B-Q4_K_M".to_string(),
-                    source_kind: None,
-                    revision: None,
-                }),
-                mmproj_ref: None,
-            }],
-            plugins: vec![],
-            config_toml: None,
-            mesh_requirements: None,
-        };
-
-        let encoded = snapshot.encode_to_vec();
-        let decoded = NodeConfigSnapshot::decode(encoded.as_slice())
-            .expect("payload without gpu_id must still decode");
-        let restored = proto_config_to_mesh(&decoded);
-
-        assert_eq!(
-            restored.gpu.assignment,
-            crate::plugin::GpuAssignment::Pinned
-        );
-        assert_eq!(restored.models.len(), 1);
-        assert_eq!(restored.models[0].gpu_id, None);
-        assert_eq!(restored.models[0].ctx_size, Some(4096));
-    }
-
-    #[test]
-    fn test_peer_announcement_first_joined_mesh_ts_roundtrip() {
-        use iroh::SecretKey;
-        use std::collections::HashMap;
-
-        let peer_id = EndpointId::from(SecretKey::from_bytes(&[0xEF; 32]).public());
-
-        let ann_with_timestamp = super::PeerAnnouncement {
-            addr: iroh::EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            role: super::NodeRole::Worker,
-            first_joined_mesh_ts: Some(1_700_000_000_000u64),
-            models: vec![],
-            vram_bytes: 0,
-            model_source: None,
-            serving_models: vec![],
-            hosted_models: None,
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec![],
-            version: None,
-            model_demand: HashMap::new(),
-            mesh_id: None,
-            mesh_policy_hash: None,
-            gpu_name: None,
-            hostname: None,
-            is_soc: None,
-            gpu_vram: None,
-            gpu_reserved_bytes: None,
-            gpu_mem_bandwidth_gbps: None,
-            gpu_compute_tflops_fp32: None,
-            gpu_compute_tflops_fp16: None,
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: None,
-            genesis_policy: None,
-            release_attestation: None,
-            direct_admission_proof: None,
-            artifact_transfer_supported: true,
-            stage_protocol_generation_supported: true,
-            stage_status_list_supported: true,
-            advertised_model_throughput: vec![],
-            latency_ms: None,
-            latency_source: None,
-            latency_age_ms: None,
-            latency_observer_id: None,
-        };
-
-        let proto_pa = local_ann_to_proto_ann(&ann_with_timestamp);
-        assert_eq!(proto_pa.first_joined_mesh_ts, Some(1_700_000_000_000u64));
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(
-            roundtripped.first_joined_mesh_ts,
-            Some(1_700_000_000_000u64)
-        );
-
-        let ann_without_timestamp = super::PeerAnnouncement {
-            addr: iroh::EndpointAddr {
-                id: peer_id,
-                addrs: Default::default(),
-            },
-            role: super::NodeRole::Worker,
-            first_joined_mesh_ts: None,
-            models: vec![],
-            vram_bytes: 0,
-            model_source: None,
-            serving_models: vec![],
-            hosted_models: None,
-            available_models: vec![],
-            requested_models: vec![],
-            explicit_model_interests: vec![],
-            version: None,
-            model_demand: HashMap::new(),
-            mesh_id: None,
-            mesh_policy_hash: None,
-            gpu_name: None,
-            hostname: None,
-            is_soc: None,
-            gpu_vram: None,
-            gpu_reserved_bytes: None,
-            gpu_mem_bandwidth_gbps: None,
-            gpu_compute_tflops_fp32: None,
-            gpu_compute_tflops_fp16: None,
-            available_model_metadata: vec![],
-            experts_summary: None,
-            available_model_sizes: HashMap::new(),
-            served_model_descriptors: vec![],
-            served_model_runtime: vec![],
-            owner_attestation: None,
-            genesis_policy: None,
-            release_attestation: None,
-            direct_admission_proof: None,
-            artifact_transfer_supported: false,
-            stage_protocol_generation_supported: false,
-            stage_status_list_supported: false,
-            advertised_model_throughput: vec![],
-            latency_ms: None,
-            latency_source: None,
-            latency_age_ms: None,
-            latency_observer_id: None,
-        };
-
-        let proto_pa = local_ann_to_proto_ann(&ann_without_timestamp);
-        assert_eq!(proto_pa.first_joined_mesh_ts, None);
-
-        let (_, roundtripped) =
-            proto_ann_to_local(&proto_pa).expect("proto_ann_to_local must succeed");
-        assert_eq!(roundtripped.first_joined_mesh_ts, None);
-    }
-}
+pub(crate) mod tests;
